Add DropDatabaseDiscoveryRuleStatementUpdater
