+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.RDLUpdater
@@ -22,6 +22,7 @@ org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterReadw
 
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.CreateDatabaseDiscoveryRuleStatementUpdater
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterDatabaseDiscoveryRuleStatementUpdater
+org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.DropDatabaseDiscoveryRuleStatementUpdater
 
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.CreateEncryptRuleStatementUpdater
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterEncryptRuleStatementUpdater
